requirements.txt
flask
python-dotenv
openai
aiohttp
requests
httpx
gunicorn
